Update package.json and package-lock.json
Updated the "package.json" and "package-lock.json" files to include the following packages:

- "eslint": "^8.39.0"
- "eslint-config-prettier": "^8.8.0"
- "eslint-config-standard": "^17.0.0"
- "eslint-plugin-import": "^2.27.5"
- "eslint-plugin-node": "^11.1.0"
- "eslint-plugin-promise": "^6.1.1"
- "eslint-plugin-react": "^7.32.2"
- "prettier": "^2.8.8"

These packages were downloaded and added to the project to support code linting and formatting using ESLint and Prettier.
These updates ensure that the project adheres to the latest code quality and formatting standards.

This commit builds upon the previous commit that added the ".eslintignore", ".eslintrc.cjs", and "prettierignore" files.